+++ data/index.js
@@ -1,15 +1,15 @@
-import Africa from './Africa.js'
-import America from './America.js'
-import Antarctica from './Antarctica.js'
-import Asia from './Asia.js'
-import Atlantic from './Atlantic.js'
-import Australia from './Australia.js'
-import Etc from './Etc.js'
-import Europe from './Europe.js'
-import Indian from './Indian.js'
-import Pacific from './Pacific.js'
+import Africa from './zones/Africa.js'
+import America from './zones/America.js'
+import Antarctica from './zones/Antarctica.js'
+import Asia from './zones/Asia.js'
+import Atlantic from './zones/Atlantic.js'
+import Australia from './zones/Australia.js'
+import Etc from './zones/Etc.js'
+import Europe from './zones/Europe.js'
+import Indian from './zones/Indian.js'
+import Pacific from './zones/Pacific.js'
 
-let zones = Object.assign({},
+const regions = [
   Africa,
   America,
   Antarctica,
@@ -20,5 +20,12 @@ let zones = Object.assign({},
   Europe,
   Indian,
   Pacific,
-)
+]
+const zones = {}
+for (const region of regions) {
+  for (const [id, record] of Object.entries(region)) {
+    if (Object.hasOwn(zones, id)) throw new Error(`Duplicate zone record: ${id}`)
+    zones[id] = record
+  }
+}
 export default zones