+++ scripts/check-dst.js
@@ -1,29 +1,8 @@
-import metas from '../src/display/metas.js'
+/* eslint-disable no-console */
+import metas from '../data/metas.js'
 // import { zones } from '../src/data/index.js'
 
-import Africa from '../data/Africa.js'
-import America from '../data/America.js'
-import Antarctica from '../data/Antarctica.js'
-import Asia from '../data/Asia.js'
-import Atlantic from '../data/Atlantic.js'
-import Australia from '../data/Australia.js'
-import Etc from '../data/Etc.js'
-import Europe from '../data/Europe.js'
-import Indian from '../data/Indian.js'
-import Pacific from '../data/Pacific.js'
-
-let zones = Object.assign({},
-  Africa,
-  America,
-  Antarctica,
-  Asia,
-  Atlantic,
-  Australia,
-  Etc,
-  Europe,
-  Indian,
-  Pacific,
-)
+import zones from '../data/index.js'
 Object.keys(zones).forEach(k => {
   let meta = zones[k].meta
   if (zones[k].offset !== metas[meta].std[1]) {