+++ scripts/filesize.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 import fs from 'fs'
 //log the size of our builds
 const stats = fs.statSync('./builds/timezone-soft.min.cjs')