+++ src/find/parseOffset.js
@@ -1,11 +1,14 @@
 const isOffset = /^([-+]?[0-9]+)h(r?s)?$/i
 const isNumber = /^([-+]?[0-9]+)$/
-const utcOffset = /utc([\-+]?[0-9]+)$/i
-const gmtOffset = /gmt([\-+]?[0-9]+)$/i
+const utcOffset = /^utc([\-+]?[0-9]+)$/i
+const gmtOffset = /^(?:etc\/)?gmt([\-+]?[0-9]+)$/i
 
 const toIana = function (num) {
   num = Number(num)
-  if (num > -13 && num < 13) {
+  if (num === 0) {
+    return 'Etc/GMT'
+  }
+  if (num >= -12 && num <= 14) {
     num = num * -1 //it's opposite!
     num = (num > 0 ? '+' : '') + num //add plus sign
     return 'Etc/GMT' + num
@@ -14,6 +17,7 @@ const toIana = function (num) {
 }
 
 const parseOffset = function (tz) {
+  tz = tz.trim()
   // '+5hrs'
   let m = tz.match(isOffset)
   if (m !== null) {